+++ RetroFE/Source/Graphics/Component/ScrollingList.h
@@ -74,7 +74,7 @@ public:
     void setPoints( std::vector<ViewInfo *> *scrollPoints, std::vector<AnimationEvents *> *tweenPoints );
     unsigned int getSelectedIndex( );
     void setSelectedIndex( unsigned int index );
-    unsigned int getSize( );
+    size_t getSize( );
     void pageUp( );
     void pageDown( );
     void letterUp( );
@@ -114,8 +114,8 @@ public:
 private:
 
     void resetTweens( Component *c, AnimationEvents *sets, ViewInfo *currentViewInfo, ViewInfo *nextViewInfo, double scrollTime );
-    unsigned int loopIncrement( unsigned int offset, unsigned int i, unsigned int size );
-    unsigned int loopDecrement( unsigned int offset, unsigned int i, unsigned int size );
+    unsigned int loopIncrement(size_t offset, size_t i, size_t size );
+    unsigned int loopDecrement(size_t offset, size_t i, size_t size );
 
     bool layoutMode_;
     bool commonMode_;
